requirements.txt
streamlit
opencv-python
python-dotenv
transformers
Pillow
requests
numpy
torch
pandas
openpyxl
